# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Agendamento API",
        "description": "API REST para gerenciamento de agendamentos com autenticação JWT, gestão de dias fechados e QR Code para validação. Esta API permite criar, listar, visualizar e cancelar agendamentos, além de gerenciar dias bloqueados e validar QR Codes.",
        "termsOfService": "http://localhost:8000/terms",
        "contact": {
            "name": "Equipe de Desenvolvimento",
            "url": "https://github.com/seu-usuario/agendamento",
            "email": "dev@agendamento.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/admin/agendamento/{uuid}": {
            "get": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Visualizar agendamento",
                "description": "Visualiza os dados de um agendamento específico pelo UUID",
                "operationId": "9adb7e74ccec2d3ac4fbda0ab28a816b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do agendamento",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "15/12/2023"
                                        },
                                        "horario": {
                                            "type": "string",
                                            "example": "14:00"
                                        },
                                        "grupo": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "quantidade": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "observacao": {
                                            "type": "string",
                                            "example": "Observação do agendamento"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Agendamento não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Agendamento não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/agendamento/{id}": {
            "delete": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Cancelar agendamento",
                "description": "Cancela um agendamento existente",
                "operationId": "b8e8969cc3ca4223a778b01e5c708547",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agendamento cancelado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Agendamento cancelado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/agendamento/vagas-por-horario": {
            "get": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Consultar vagas por horário",
                "description": "Consulta a disponibilidade de vagas em uma data específica",
                "operationId": "6fcf6c0012911d6b89c23b7882df1dd1",
                "parameters": [
                    {
                        "name": "data",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-12-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Disponibilidade de vagas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "2023-12-15"
                                        },
                                        "bloqueado": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "motivo": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "horarios": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "hora": {
                                                        "type": "string",
                                                        "example": "09:00"
                                                    },
                                                    "vagas": {
                                                        "type": "integer",
                                                        "example": 48
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Digite o token Bearer no formato: Bearer {seu-token-jwt}"
            }
        }
    },
...
}